+++ datumaro/plugins/data_formats/datumaro/base.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022 Intel Corporation
+# Copyright (C) 2023 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
@@ -21,10 +21,8 @@ from datumaro.components.annotation import (
 )
 from datumaro.components.dataset_base import DatasetItem, SubsetBase
 from datumaro.components.errors import DatasetImportError
-from datumaro.components.format_detection import FormatDetectionContext
-from datumaro.components.importer import Importer
 from datumaro.components.media import Image, MediaElement, PointCloud
-from datumaro.util import parse_json, parse_json_file
+from datumaro.util import parse_json_file
 
 from .format import DatumaroPath
 
@@ -281,21 +279,3 @@ class DatumaroBase(SubsetBase):
                 raise NotImplementedError()
 
         return loaded
-
-
-class DatumaroImporter(Importer):
-    @classmethod
-    def detect(cls, context: FormatDetectionContext) -> None:
-        annot_file = context.require_file("annotations/*.json")
-
-        with context.probe_text_file(
-            annot_file,
-            'must be a JSON object with "categories" ' 'and "items" keys',
-        ) as f:
-            contents = parse_json(f.read())
-            if not {"categories", "items"} <= contents.keys():
-                raise Exception
-
-    @classmethod
-    def find_sources(cls, path):
-        return cls._find_sources_recursive(path, ".json", "datumaro", dirname="annotations")