+++ datumaro/plugins/data_formats/datumaro_binary/base.py
@@ -1,0 +1,9 @@
+# Copyright (C) 2023 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from ..datumaro.base import DatumaroBase
+
+
+class DatumaroBinaryBase(DatumaroBase):
+    """"""