+++ datumaro/plugins/data_formats/datumaro_binary/format.py
@@ -3,13 +3,14 @@
 # SPDX-License-Identifier: MIT
 
 
-class DatumaroPath:
+class DatumaroBinaryPath:
     IMAGES_DIR = "images"
     ANNOTATIONS_DIR = "annotations"
     PCD_DIR = "point_clouds"
     RELATED_IMAGES_DIR = "related_images"
     MASKS_DIR = "masks"
 
-    ANNOTATION_EXT = ".json"
+    ANNOTATION_EXT = ".datumaro"
     IMAGE_EXT = ".jpg"
     MASK_EXT = ".png"
+    SIGNATURE = "signature:datumaro_binary"