+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
+### Added
 - Add with_subset_dirs decorator
   (<https://github.com/openvinotoolkit/datumaro/pull/816>)
+
+### Changed
+- Refactor Datumaro format code and test code
+  (<https://github.com/openvinotoolkit/datumaro/pull/824>)
 
 ## 24/02/2023 - Release v1.0.0
 ### Added